CleanUp Views and Urls
