--- base/urls.py
@@ -1,16 +1,0 @@
-from django.urls import path
-from . import views
-
-from rest_framework_simplejwt.views import (
-    TokenObtainPairView,
-)
-
-urlpatterns = [
-    path('users/token/', views.MyTokenObtainPairView.as_view(), name='token_obtain_pair'),
-    path('', views.getRoutes, name='routes'),
-    path('users/profile/', views.getUserProfile, name='users-profile'),
-    path('users/register/', views.registerUser, name='users-register'),
-    path('users/', views.getUsers, name='users'),
-    path('products/', views.getProducts, name='products'),
-    path('products/<str:pk>', views.getProduct, name='product'),
-]
