+++ base/urls/product_urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from base.views import product_views as views
+
+urlpatterns = [
+    path('', views.getProducts, name='products'),
+    path('<str:pk>', views.getProduct, name='product-detail'),
+]
