+++ base/views/product_views.py
@@ -1,0 +1,25 @@
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated, IsAdminUser
+from rest_framework.response import Response
+
+
+from base.models import Product
+
+from base.serializers import ProductSerializer
+
+from django.contrib.auth.hashers import make_password
+from rest_framework import status
+
+
+@api_view(['GET'])
+def getProducts(request):
+    products = Product.objects.all()
+    serializer = ProductSerializer(products, many=True)
+    return Response(serializer.data)
+    
+@api_view(['GET'])
+def getProduct(request, pk):
+    product = Product.objects.get(pk=pk)
+    serializer = ProductSerializer(product, many=False)
+
+    return Response(serializer.data)
