+++ base/views/user_views.py
@@ -1,14 +1,10 @@
-from email import message
-from django.http import JsonResponse
-
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated, IsAdminUser
 from rest_framework.response import Response
 from base import serializers
 
-from base.serializers import ProductSerializer, UserSerializer, UserSerializerWithToken
+from base.serializers import UserSerializer, UserSerializerWithToken
 
-from .models import Product
 from django.contrib.auth.models import User
 
 from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
@@ -30,30 +26,6 @@ class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
 
 class MyTokenObtainPairView(TokenObtainPairView):
     serializer_class = MyTokenObtainPairSerializer
-
-@api_view(['GET'])
-def getRoutes(request):
-    routes = [
-        '/api/products/',
-        '/api/products/create/',
-        
-        '/api/products/upload/',
-        
-        '/api/products/<id>/reviews/',
-        
-        '/api/products/top/',
-        '/api/products/<id>/',
-        
-        '/api/products/delete/<id>/',
-        '/api/products/<update>/<id>/',
-
-        '/api/users/profile/',
-        '/api/users/register/',
-        '/api/users/token/',
-        '/api/users/'
-    ]
-    return Response(routes)
-
 
 @api_view(['POST'])
 def registerUser(request):
@@ -84,17 +56,4 @@ def getUserProfile(request):
 def getUsers(request):
     users = User.objects.all()
     serializer = UserSerializer(users, many=True)
-    return Response(serializer.data)
-
-@api_view(['GET'])
-def getProducts(request):
-    products = Product.objects.all()
-    serializer = ProductSerializer(products, many=True)
-    return Response(serializer.data)
-    
-@api_view(['GET'])
-def getProduct(request, pk):
-    product = Product.objects.get(pk=pk)
-    serializer = ProductSerializer(product, many=False)
-
     return Response(serializer.data)
